+++ test/request.test.js
@@ -1,22 +1,19 @@
 var nodeio = require('node.io'),
-    processor = new nodeio.Processor(),
     http = require('http'),
-    JobClass = nodeio.JobClass,
     assert = require('assert');
     
-var job = new JobClass();
+var port = 24510, timeout = 2000;
 
-var i = 24510;
-
-//Why do these tests fail on linux?! D:
-
-job.debug = function () {};
-
-//Throw a warning on ECONNREFUSED rather than fail the entire test suite
-job.fail = function (input, status) {
-    if (status === 'ECONNREFUSED') {
-        console.log('\x1B[33mWARNING\x1B[0m: \x1B[31mECONNREFUSED\x1B[0m (see request.test.js)');
-    }
+function createJob() {
+    var JobClass = nodeio.JobClass, job = new JobClass();
+    job.debug = function () {};
+    //Throw a warning on ECONNREFUSED rather than fail the entire test suite
+    job.fail = function (input, status) {
+        if (status === 'ECONNREFUSED') {
+            console.log('\x1B[33mWARNING\x1B[0m: \x1B[31mECONNREFUSED\x1B[0m (see request.test.js)');
+        }
+    };
+    return job;
 }
 
 function close (server) {
@@ -29,14 +26,17 @@ module.exports = {
     
     'test GET request': function() {
     
+        var job = createJob();
+    
         var server = http.createServer(function (req, res) {
-            res.writeHead(200,{'Content-Type': 'text/plain'});
+            res.writeHead(200, {'Content-Type': 'text/plain'});
             res.end('Hello World');
         });
         
-        server.listen(++i);
-                
-        job.get('http://127.0.0.1:'+i+'/', function(err, data, headers) {
+        server.listen(++port);
+        
+        job.get('http://127.0.0.1:'+port+'/', function(err, data, headers) {
+            if (err) throw err;
             assert.equal('text/plain', headers['content-type']);
             assert.equal('Hello World', data);
             close(server);
@@ -44,11 +44,13 @@ module.exports = {
         
         setTimeout(function() {
             close(server);
-        }, 1000);
+        }, timeout);
     },
     
     'test GET request with custom headers': function() {
     
+        var job = createJob();
+        
         var server = http.createServer(function (req, res) {
             if (req.headers.foo === 'bar') {
                 res.writeHead(200,{'Content-Type': 'text/plain'});
@@ -58,9 +60,9 @@ module.exports = {
             }
         });
         
-        server.listen(++i);
+        server.listen(++port);
                 
-        job.get('http://127.0.0.1:'+i+'/', {foo:'bar'}, function(err, data, headers) {
+        job.get('http://127.0.0.1:'+port+'/', {foo:'bar'}, function(err, data, headers) {
             assert.equal('text/plain', headers['content-type']);
             assert.equal('Headers ok', data);
             close(server);
@@ -68,23 +70,25 @@ module.exports = {
         
         setTimeout(function() {
             close(server);
-        }, 1000);
+        }, timeout);
     },
     
     'test GET request with pre-parse callback': function() {
     
+        var job = createJob();
+        
         var server = http.createServer(function (req, res) {
             res.writeHead(200,{'Content-Type': 'text/plain'});
             res.end('&gt;&lt;');
         });
         
-        server.listen(++i);
+        server.listen(++port);
         
         var parse = function(str) {
             return str.replace('&gt;','>').replace('&lt;','<');
         }
         
-        job.get('http://127.0.0.1:'+i+'/', function(err, data, headers) {
+        job.get('http://127.0.0.1:'+port+'/', function(err, data, headers) {
             assert.equal('text/plain', headers['content-type']);
             assert.equal('><', data);
             close(server);
@@ -92,11 +96,13 @@ module.exports = {
         
         setTimeout(function() {
             close(server);
-        }, 1000);
+        }, timeout);
     },
     
     'test POST request': function() {
     
+        var job = createJob();
+        
         var server = http.createServer(function (req, res) {
             var data = '';
             req.setEncoding('utf8');
@@ -111,9 +117,9 @@ module.exports = {
             });
         });
         
-        server.listen(++i);
+        server.listen(++port);
                 
-        job.post('http://127.0.0.1:'+i+'/', {foo:'bar'}, function(err, data, headers) {
+        job.post('http://127.0.0.1:'+port+'/', {foo:'bar'}, function(err, data, headers) {
             assert.equal('text/plain', headers['content-type']);
             assert.equal('Post ok', data);
             close(server);
@@ -121,19 +127,21 @@ module.exports = {
         
         setTimeout(function() {
             close(server);
-        }, 1000);
+        }, timeout);
     },
     
     'test GET request returning the dom': function() {
     
+        var job = createJob();
+        
         var server = http.createServer(function (req, res) {
             res.writeHead(200,{'Content-Type': 'text/plain'});
             res.end('<p class="a"></p>');
         });
         
-        server.listen(++i);
+        server.listen(++port);
                 
-        job.getHtml('http://127.0.0.1:'+i+'/', function(err, $, data, headers) {
+        job.getHtml('http://127.0.0.1:'+port+'/', function(err, $, data, headers) {
             assert.equal('text/plain', headers['content-type']);
             assert.equal('<p class="a"></p>', data);
             assert.equal('a', $('p').attribs['class']);
@@ -142,11 +150,13 @@ module.exports = {
         
         setTimeout(function() {
             close(server);
-        }, 1000);
+        }, timeout);
     },
     
     'test POST request returning the dom': function() {
     
+        var job = createJob();
+        
         var server = http.createServer(function (req, res) {
             var data = '';
             req.setEncoding('utf8');
@@ -161,9 +171,9 @@ module.exports = {
             });
         });
         
-        server.listen(++i);
+        server.listen(++port);
                 
-        job.postHtml('http://127.0.0.1:'+i+'/', {foo:'bar'}, function(err, $, data, headers) {
+        job.postHtml('http://127.0.0.1:'+port+'/', {foo:'bar'}, function(err, $, data, headers) {
             assert.equal('text/plain', headers['content-type']);
             assert.equal('<p class="a"></p>', data);
             assert.equal('a', $('p').attribs['class']);
@@ -172,7 +182,97 @@ module.exports = {
         
         setTimeout(function() {
             close(server);
-        }, 1000);
-    }   
+        }, timeout);
+    },
     
+    'test nested request': function() {
+    
+        var p = ++port, job = createJob();
+        
+        var server = http.createServer(function (req, res) {
+            if (!req.headers.cookie) {
+                res.writeHead(200, {'Content-Type': 'text/plain', 'Set-Cookie': 'foo=bar'});
+                res.end('Ok');
+            } else if (req.headers.cookie === 'foo=bar' && req.headers.referer === 'http://127.0.0.1:'+p+'/') {
+                res.writeHead(200, {'Content-Type': 'text/plain'});
+                res.end('Ok2');
+            } else {
+                res.end();
+            }
+        });
+                
+        server.listen(p);
+        
+        job.get('http://127.0.0.1:'+p+'/', function(err, data, headers) {
+            assert.equal('Ok', data);
+            job.get('http://127.0.0.1:'+p+'/', function(err, data, headers) {
+                assert.equal('Ok2', data);
+                close(server);
+            });
+        });
+        
+        setTimeout(function() {
+            close(server);
+        }, timeout);
+    },
+    
+    'test GET request with custom headers 2': function() {
+    
+        var job = createJob();
+        
+        var server = http.createServer(function (req, res) {
+            if (req.headers.foo === 'bar' && req.headers.cookie === 'coo=kie' && req.headers['user-agent'] === 'Firefox') {
+                res.writeHead(200, {'Content-Type': 'text/plain'});
+                res.end('Headers ok');
+            } else {
+                res.end();
+            }
+        });
+        
+        server.listen(++port);
+        
+        job.setHeader('foo', 'bar');
+        job.setCookie('coo', 'kie');
+        job.setUserAgent('Firefox');
+        
+        job.get('http://127.0.0.1:'+port+'/', function(err, data, headers) {
+            assert.equal('text/plain', headers['content-type']);
+            assert.equal('Headers ok', data);
+            close(server);
+        });
+        
+        setTimeout(function() {
+            close(server);
+        }, timeout);
+    },
+    
+    'test GET request with addCookie': function() {
+    
+        var job = createJob();
+        
+        var server = http.createServer(function (req, res) {
+            var cookies = req.headers.cookie.split('; ');
+            if (cookies[0] === 'coo=kie' && cookies[1] === 'foo=bar') {
+                res.writeHead(200, {'Content-Type': 'text/plain'});
+                res.end('Headers ok');
+            } else {
+                res.end();
+            }
+        });
+        
+        server.listen(++port);
+        
+        job.addCookie('coo', 'kie');
+        job.addCookie('foo', 'bar');
+        
+        job.get('http://127.0.0.1:'+port+'/', function(err, data, headers) {
+            assert.equal('text/plain', headers['content-type']);
+            assert.equal('Headers ok', data);
+            close(server);
+        });
+        
+        setTimeout(function() {
+            close(server);
+        }, timeout);
+    },
 }